changing file and component name
