+++ src/components/SizeSelector/index.scss
